+++ dataminer.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-# rubocop:disable Metrics/ClassLength
 # rubocop:disable Metrics/BlockLength
 
 require 'bundler'
 Bundler.require(:default, ENV.fetch('RACK_ENV', 'development'))
 
+require 'base64'
 require 'pstore'
 require './lib/types_for_dry'
 require './lib/crossbeams_responses'
@@ -23,7 +23,7 @@ ENV['ROOT'] = File.dirname(__FILE__)
 ENV['VERSION'] = File.read('VERSION')
 # ENV['REPORTS_LOCATION'] ||= File.expand_path('../../../roda_frame/reports', __FILE__)
 # ENV['REPORTS_LOCATION'] ||= File.expand_path('../../label_designer/grid_definitions/dataminer_queries', __FILE__)
-ENV['GRID_QUERIES_LOCATION'] ||= File.expand_path('../../label_designer/grid_definitions/dataminer_queries', __FILE__)
+ENV['GRID_QUERIES_LOCATION'] ||= File.expand_path('../label_designer/grid_definitions/dataminer_queries', __dir__)
 
 DM_CONNECTIONS = DataminerConnections.new
 
@@ -87,6 +87,25 @@ class Dataminer < Roda
     r.assets unless ENV['RACK_ENV'] == 'production'
     r.public
 
+    # Routes that must work without authentication
+    # --------------------------------------------
+    r.on 'webquery', String do |id|
+      # A dummy user
+      user = DevelopmentApp::User.new(id: 0, login_name: 'webquery', user_name: 'webquery', password_hash: 'dummy', email: nil, active: true)
+      interactor = DataminerApp::PreparedReportInteractor.new(user, {}, { route_url: request.path }, {})
+      interactor.prepared_report_as_html(id)
+    end
+
+    # https://support.office.com/en-us/article/import-data-from-database-using-native-database-query-power-query-f4f448ac-70d5-445b-a6ba-302db47a1b00?ui=en-US&rs=en-US&ad=US
+    r.on 'xmlreport', String do |id|
+      # A dummy user
+      user = DevelopmentApp::User.new(id: 0, login_name: 'webquery', user_name: 'webquery', password_hash: 'dummy', email: nil, active: true)
+      interactor = DataminerApp::PreparedReportInteractor.new(user, {}, { route_url: request.path }, {})
+      interactor.prepared_report_as_xml(id)
+    end
+    # Do the same as XML?
+    # --------------------------------------------
+
     r.rodauth
     rodauth.require_authentication
     r.redirect('/login') if current_user.nil? # Session might have the incorrect user_id
@@ -133,97 +152,97 @@ class Dataminer < Roda
       view(inline: '<div class="crossbeams-error-note"><strong>Error</strong><br>The requested resource was not found.</div>')
     end
 
-    # Generic grid lists.
-    r.on 'list' do
-      r.on :id do |id|
-        r.is do
-          session[:last_grid_url] = "/list/#{id}"
-          show_page { render_data_grid_page(id) }
-        end
-
-        r.on 'with_params' do
-          if fetch?(r)
-            show_partial { render_data_grid_page(id, query_string: request.query_string) }
-          else
-            session[:last_grid_url] = "/list/#{id}/with_params?#{request.query_string}"
-            show_page { render_data_grid_page(id, query_string: request.query_string) }
-          end
-        end
-
-        r.on 'multi' do
-          if fetch?(r)
-            show_partial { render_data_grid_page_multiselect(id, params) }
-          else
-            show_page { render_data_grid_page_multiselect(id, params) }
-          end
-        end
-
-        r.on 'grid' do
-          response['Content-Type'] = 'application/json'
-          begin
-            if params && !params.empty?
-              render_data_grid_rows(id, ->(program, permission) { auth_blocked?(program, permission) }, params)
-            else
-              render_data_grid_rows(id, ->(program, permission) { auth_blocked?(program, permission) })
-            end
-          rescue StandardError => e
-            show_json_exception(e)
-          end
-        end
-
-        r.on 'grid_multi', String do |key|
-          response['Content-Type'] = 'application/json'
-          begin
-            render_data_grid_multiselect_rows(id, ->(program, permission) { auth_blocked?(program, permission) }, key, params)
-          rescue StandardError => e
-            show_json_exception(e)
-          end
-        end
-      end
-    end
-
-    r.on 'print_grid' do
-      @layout = Crossbeams::Layout::Page.build(grid_url: params[:grid_url]) do |page, _|
-        page.add_grid('crossbeamsPrintGrid', params[:grid_url], caption: 'Print', for_print: true)
-      end
-      view('crossbeams_layout_page', layout: 'print_layout')
-    end
-
-    # Generic code for grid searches.
-    r.on 'search' do
-      r.on :id do |id|
-        r.is do
-          render_search_filter(id, params)
-        end
-
-        r.on 'run' do
-          session[:last_grid_url] = "/search/#{id}?rerun=y"
-          show_page { render_search_grid_page(id, params) }
-        end
-
-        r.on 'grid' do
-          response['Content-Type'] = 'application/json'
-          render_search_grid_rows(id, params, ->(program, permission) { auth_blocked?(program, permission) })
-        end
-
-        r.on 'xls' do
-          caption, xls = render_excel_rows(id, params)
-          response.headers['content_type'] = 'application/vnd.ms-excel'
-          response.headers['Content-Disposition'] = "attachment; filename=\"#{caption.strip.gsub(%r{[/:*?"\\<>\|\r\n]}i, '-') + '.xls'}\""
-          response.write(xls) # NOTE: could this use streaming to start downloading quicker?
-        rescue Sequel::DatabaseError => e
-          view(inline: <<-HTML)
-          <p style='color:red;'>There is a problem with the SQL definition of this report:</p>
-          <p>Report: <em>#{caption}</em></p>The error message is:
-          <pre>#{e.message}</pre>
-          <button class="pure-button" onclick="crossbeamsUtils.toggleVisibility('sql_code', this);return false">
-            <i class="fa fa-info"></i> Toggle SQL
-          </button>
-          <pre id="sql_code" style="display:none;"><%= sql_to_highlight(@rpt.runnable_sql) %></pre>
-          HTML
-        end
-      end
-    end
+    # # Generic grid lists.
+    # r.on 'list' do
+    #   r.on :id do |id|
+    #     r.is do
+    #       session[:last_grid_url] = "/list/#{id}"
+    #       show_page { render_data_grid_page(id) }
+    #     end
+    #
+    #     r.on 'with_params' do
+    #       if fetch?(r)
+    #         show_partial { render_data_grid_page(id, query_string: request.query_string) }
+    #       else
+    #         session[:last_grid_url] = "/list/#{id}/with_params?#{request.query_string}"
+    #         show_page { render_data_grid_page(id, query_string: request.query_string) }
+    #       end
+    #     end
+    #
+    #     r.on 'multi' do
+    #       if fetch?(r)
+    #         show_partial { render_data_grid_page_multiselect(id, params) }
+    #       else
+    #         show_page { render_data_grid_page_multiselect(id, params) }
+    #       end
+    #     end
+    #
+    #     r.on 'grid' do
+    #       response['Content-Type'] = 'application/json'
+    #       begin
+    #         if params && !params.empty?
+    #           render_data_grid_rows(id, ->(program, permission) { auth_blocked?(program, permission) }, params)
+    #         else
+    #           render_data_grid_rows(id, ->(program, permission) { auth_blocked?(program, permission) })
+    #         end
+    #       rescue StandardError => e
+    #         show_json_exception(e)
+    #       end
+    #     end
+    #
+    #     r.on 'grid_multi', String do |key|
+    #       response['Content-Type'] = 'application/json'
+    #       begin
+    #         render_data_grid_multiselect_rows(id, ->(program, permission) { auth_blocked?(program, permission) }, key, params)
+    #       rescue StandardError => e
+    #         show_json_exception(e)
+    #       end
+    #     end
+    #   end
+    # end
+    #
+    # r.on 'print_grid' do
+    #   @layout = Crossbeams::Layout::Page.build(grid_url: params[:grid_url]) do |page, _|
+    #     page.add_grid('crossbeamsPrintGrid', params[:grid_url], caption: 'Print', for_print: true)
+    #   end
+    #   view('crossbeams_layout_page', layout: 'print_layout')
+    # end
+    #
+    # # Generic code for grid searches.
+    # r.on 'search' do
+    #   r.on :id do |id|
+    #     r.is do
+    #       render_search_filter(id, params)
+    #     end
+    #
+    #     r.on 'run' do
+    #       session[:last_grid_url] = "/search/#{id}?rerun=y"
+    #       show_page { render_search_grid_page(id, params) }
+    #     end
+    #
+    #     r.on 'grid' do
+    #       response['Content-Type'] = 'application/json'
+    #       render_search_grid_rows(id, params, ->(program, permission) { auth_blocked?(program, permission) })
+    #     end
+    #
+    #     r.on 'xls' do
+    #       caption, xls = render_excel_rows(id, params)
+    #       response.headers['content_type'] = 'application/vnd.ms-excel'
+    #       response.headers['Content-Disposition'] = "attachment; filename=\"#{caption.strip.gsub(%r{[/:*?"\\<>\|\r\n]}i, '-') + '.xls'}\""
+    #       response.write(xls) # NOTE: could this use streaming to start downloading quicker?
+    #     rescue Sequel::DatabaseError => e
+    #       view(inline: <<-HTML)
+    #       <p style='color:red;'>There is a problem with the SQL definition of this report:</p>
+    #       <p>Report: <em>#{caption}</em></p>The error message is:
+    #       <pre>#{e.message}</pre>
+    #       <button class="pure-button" onclick="crossbeamsUtils.toggleVisibility('sql_code', this);return false">
+    #         <i class="fa fa-info"></i> Toggle SQL
+    #       </button>
+    #       <pre id="sql_code" style="display:none;"><%= sql_to_highlight(@rpt.runnable_sql) %></pre>
+    #       HTML
+    #     end
+    #   end
+    # end
   end
 end
 # rubocop:enable Metrics/ClassLength